docker_files/requirements_selenium.txt
webdriver_manager
selenium
pytz
google-cloud-storage
google-cloud-logging
protobuf==3.20.1